Update workflow name and cron schedule
The project was originally misnamed "bkb-website-backend-tests". However, this isn't a backend only for the website, but for our entire activity, so renaming it accordingly and for consistency.

Also, updating the cron schedule to run only once at 6AM, as we now have a cache system in place on the booking views, which is invalidated and refreshed at 3AM CET. So we change the schedule to only run once per day at 6AM, when the data should still be fresh and up to date. If a more frequent check is required, using the new "no-cache" view would be better.